+++ optaplanner-examples/src/main/resources/org/optaplanner/examples/tennis/optional/score/drools/tennisConstraints.drl
@@ -25,8 +25,8 @@ import org.optaplanner.examples.tennis.domain.UnavailabilityPenalty;
 import org.optaplanner.examples.tennis.domain.TeamAssignment;
 
 import org.apache.commons.lang3.tuple.Pair;
-import accumulate org.optaplanner.examples.common.optional.score.LoadBalanceByCountAccumulateFunction loadBalanceByCount;
-import accumulate org.optaplanner.examples.common.optional.score.LoadBalanceAccumulateFunction loadBalance;
+import accumulate org.optaplanner.examples.common.optional.score.drools.LoadBalanceByCountAccumulateFunction loadBalanceByCount;
+import accumulate org.optaplanner.examples.common.optional.score.drools.LoadBalanceAccumulateFunction loadBalance;
 
 global HardMediumSoftScoreHolder scoreHolder;
