+++ optaplanner-examples/src/test/java/org/optaplanner/examples/common/optional/score/drools/LoadBalanceByCountAccumulateFunctionTest.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.common.optional.score;
+package org.optaplanner.examples.common.optional.score.drools;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import org.optaplanner.examples.common.optional.score.LoadBalanceByCountAccumulateFunction;
+import org.optaplanner.examples.common.optional.score.drools.LoadBalanceByCountAccumulateFunction;
 
 public class LoadBalanceByCountAccumulateFunctionTest {
