+++ optaplanner-examples/src/test/java/org/optaplanner/examples/conferencescheduling/optional/score/drools/ConferenceSchedulingConstraintsXlsxTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.conferencescheduling.solver;
+package org.optaplanner.examples.conferencescheduling.optional.score.drools;
 
 import static org.optaplanner.examples.common.persistence.AbstractXlsxSolutionFileIO.DAY_FORMATTER;
 import static org.optaplanner.examples.common.persistence.AbstractXlsxSolutionFileIO.TIME_FORMATTER;
