+++ optaplanner-examples/src/test/java/org/optaplanner/examples/conferencescheduling/optional/score/drools/ConferenceSchedulingScoreSoftConstraintTest.java
@@ -1,4 +1,4 @@
-package org.optaplanner.examples.conferencescheduling.solver;
+package org.optaplanner.examples.conferencescheduling.optional.score.drools;
 
 import static org.optaplanner.examples.conferencescheduling.domain.ConferenceConstraintConfiguration.LANGUAGE_DIVERSITY;
 import static org.optaplanner.examples.conferencescheduling.domain.ConferenceConstraintConfiguration.SECTOR_CONFLICT;
